adding method to StudentDeskGroup
